Rename directories.
