+++ parser/lib/jsonapi/parser/attributes.rb
