+++ parser/lib/jsonapi/parser/error.rb
