+++ parser/lib/jsonapi/parser/resource.rb
