+++ parser/lib/jsonapi/parser/resource_identifier.rb
