+++ renderer/spec/include_directive_spec.rb
